Hardcode test suite helpers namespace
Codeception is really bad when it comes to namespacing the classes
it creates and/or requires.

The automatically generated helpers in `tests/{suitename}/` share
the same namespace (which defaults to null) with the ones created
(but empty) for the developer in `app/TestSuite/Codeception`. This
obviously breaks PSR4 autoloading which is the reason behind the
changes herein.