Prefer `post-install-cmd`
It turns out the `post-autoload-dump` was a bad idea since the
`Cake\Codeception\Installer` class does not exist when installing
with `--no-dev` (i.e. deployment).

By pluging into the `App\Console\Installer` instead, we make
sure that the class exists before trying to call it (which
composer does not do when running `scripts`).